requirements.txt
absl-py==2.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.2
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.10.0
appdirs==1.4.4
apriori==1.0.0
apyori==1.1.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
automata-lib==5.0.0
azure-core==1.30.1
azure-functions==1.21.3
azure-storage-blob==12.19.1
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bard==0.1
beautifulsoup4==4.12.3
black==22.1.0
bleach==4.1.0
boto3==1.29.5
botocore==1.32.5
bs4==0.0.1
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.1.2
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.1
cryptography==42.0.5
cycler==0.11.0
dataclasses-json==0.6.7
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
dm-tree==0.1.8
dnspython==2.3.0
docstring_parser==0.16
einops==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==0.8.2
fastapi==0.116.1
fastjsonschema==2.19.1
filelock==3.12.4
filetype==1.2.0
fire==0.4.0
Flask==2.1.1
Flask-Cors==3.0.10
flatbuffers==23.5.26
fonttools==4.32.0
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.2
gast==0.5.4
google-ai-generativelanguage==0.6.18
google-api-core==2.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.56.0
google-generativeai==0.3.2
google-pasta==0.2.0
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
gradio_client==0.11.0
graphviz==0.19.2
greenlet==3.0.0
groq==0.31.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.3.0
h5py==3.10.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.29.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jiter==0.4.2
jmespath==1.0.1
joblib==1.5.1
jovian==0.2.47
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.5.12
kiwisolver==1.4.2
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-experimental==0.0.49
langchain-google-genai==2.1.10
langchain-google-vertexai==1.0.5
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-mongodb==0.7.0
langchain-openai==0.0.5
langchain-pinecone==0.0.3
langchain-qdrant==0.2.0
langchain-tavily==0.2.11
langchain-text-splitters==0.3.10
langchainhub==0.1.14
langdetect==1.0.9
langgraph==0.0.21
langsmith==0.4.23
lark==1.2.2
libclang==16.0.6
loguru==0.7.2
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistralai==0.0.8
mistune==3.0.2
ml-dtypes==0.2.0
monotonic==1.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
munkres==1.1.4
mypy-extensions==0.4.3
namex==0.0.7
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.4
networkx==2.7.1
nltk==3.7
notebook==7.0.7
notebook-as-pdf==0.5.0
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.1
openai==1.11.0
opendatasets==0.1.22
openpyxl==3.1.2
opt-einsum==3.3.0
optimum==1.17.1
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.6.0
packaging==23.2
palettable==3.3.3
pandas==2.2.0
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
peewee==3.17.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.1.0
pinecone-client==3.1.0
pip==25.2
platformdirs==2.5.0
plumbum==1.8.1
ply==3.11
portalocker==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.43
propcache==0.3.2
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.8
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydot==1.4.2
pydotplus==2.0.2
pydub==0.25.1
pyee==8.2.2
pyensae==1.3.884
pyfpgrowth==1.0
pygame==2.3.0
Pygments==2.16.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
pymongo==4.14.1
pyparsing==3.0.9
pypdf==6.0.0
PyPDF2==3.0.1
PyPika==0.48.9
pyppeteer==1.0.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
pyquickhelper==1.11.3731
pyquicksetup==0.2
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-slugify==8.0.0
pytz==2022.7
pyu2f==0.1.5
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.15.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.8.8
requests==2.32.5
requests-futures==1.0.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.7.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.11.1
seaborn==0.12.2
selenium==4.17.2
Send2Trash==1.8.2
sentence-transformers==5.1.0
sentencepiece==0.2.0
setuptools==68.0.0
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0
slr-parser==0.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.1.4
starlette==0.47.3
stringcase==1.2.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.21.4
tornado==6.4
tqdm==4.66.1
tradingview-ta==3.3.0
traitlets==5.14.1
transformers==4.55.2
trio==0.24.0
trio-websocket==0.11.1
types-protobuf==4.24.0.20240129
types-python-dateutil==2.8.19.20240106
types-pytz==2022.4.0.0
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing-utils==0.1.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uuid==1.30
uvicorn==0.35.0
vboxapi==1.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.4
Werkzeug==2.1.1
wheel==0.38.4
widgetsnbextension==4.0.9
win32-setctime==1.1.0
windows-curses==2.3.0
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.4.1
yahooquery==2.3.7
yarl==1.20.1
yfinance==0.2.36
zipp==3.11.0
zstandard==0.24.0
